连接下载逻辑

+++ PixivDownloaderX/Utilities/ArtworkIdValidator.cs
@@ -6,7 +6,7 @@ public class ArtworkIdValidator : ValidationAttribute
 {
     protected override ValidationResult? IsValid(object? value, ValidationContext validationContext)
     {
-        return uint.TryParse(value?.ToString(), out _)
+        return ulong.TryParse(value?.ToString(), out _)
             ? ValidationResult.Success
             : new ValidationResult("Invalid Artwork Id");
     }
